+++ digital_archive/console.py
@@ -8,6 +8,8 @@ End with a blank line before the last triple quote.
 # Imports
 # -----------------------------------------------------------------------------
 import click
+from .utils import path_utils
+from .identify import reports
 
 # -----------------------------------------------------------------------------
 # Function Definitions
@@ -16,10 +18,12 @@ import click
 
 @click.group()
 def cli():
-    """A simple Command Line Tool."""
+    """Command Line Tool for handling Aarhus Digital Archive handins."""
+    pass
 
 
-@cli.command("hello")
-def hello():
-    click.confirm("Do you want to continue?", abort=True)
-    click.echo("Hi there.")
+@cli.command()
+@click.option("--path", type=click.Path(exists=True, file_okay=False))
+def report(path):
+    file_exts, empty_dirs = path_utils.explore_dir(path)
+    reports.report_results(file_exts, empty_dirs, path)
